ButtonFix

Approved-by: Дмитро Ламашевський <d.lomashevsky@massaccess.net>

+++ bcsapp/ViewModels/AplicationViewModel.cs
@@ -140,6 +140,7 @@ namespace bcsapp.ViewModels
          
         private void RemoveRoleToUser(ResponseRoleClass obj)
         {
+            if (obj == null) return;
             bool isDone;
             TransactionService.RemoveRoleToUser(new Transaction(new ServerLib.JTypes.Client.RequestUserRoleDeleteClass()
             {
